% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_eb.R
\name{scatter_eb}
\alias{scatter_eb}
\title{Add error bars to a scatter plot}
\usage{
scatter_eb(Xval, Yval, sdval, eb.width, eb.length, ...)
}
\arguments{
\item{Xval}{x values}

\item{Yval}{y values}

\item{sdval}{length of error bar along the X axis (ie. standard deviation or standard error of the mean)}

\item{eb.width}{width of error bar}

\item{eb.length}{length of error bar}
}
\value{
error bars
}
\description{
Add error bars to a scatter plot
}
